PM-5307 - move metadata, rename status enum
