+++ src/api/metadata/metadata.module.ts
@@ -11,8 +11,6 @@ import { ProductTemplateModule } from './product-template/product-template.modul
 import { ProjectTemplateModule } from './project-template/project-template.module';
 import { ProjectTypeModule } from './project-type/project-type.module';
 import { WorkManagementPermissionModule } from './work-management-permission/work-management-permission.module';
-import { ProjectPostIndustryModule } from './project-post-industry/project-post-industry.module';
-import { ProjectPostCategoryModule } from './project-post-category/project-post-category.module';
 
 /**
  * Aggregates all metadata sub-modules used by the projects API.
@@ -46,8 +44,6 @@ import { ProjectPostCategoryModule } from './project-post-category/project-post-
     PlanConfigModule,
     PriceConfigModule,
     WorkManagementPermissionModule,
-    ProjectPostIndustryModule,
-    ProjectPostCategoryModule,
   ],
   controllers: [MetadataListController],
   providers: [MetadataListService],
