+++ src/api/project-showcase-post/project-post-industry/project-post-industry.service.ts
@@ -4,60 +4,60 @@ import {
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
-import { Prisma, ProjectPostCategory } from '@prisma/client';
+import { Prisma, ProjectPostIndustry } from '@prisma/client';
 import { PrismaErrorService } from 'src/shared/modules/global/prisma-error.service';
 import { PrismaService } from 'src/shared/modules/global/prisma.service';
 import { EventBusService } from 'src/shared/modules/global/eventBus.service';
 import {
   PROJECT_METADATA_RESOURCE,
   publishMetadataEvent,
-} from '../utils/metadata-event.utils';
-import { toSerializable } from '../utils/metadata-utils';
-import { CreateProjectPostCategoryDto } from './dto/create-project-post-category.dto';
-import { ProjectPostCategoryResponseDto } from './dto/project-post-category-response.dto';
-import { UpdateProjectPostCategoryDto } from './dto/update-project-post-category.dto';
+} from 'src/api/metadata/utils/metadata-event.utils';
+import { toSerializable } from 'src/api/metadata/utils/metadata-utils';
+import { CreateProjectPostIndustryDto } from './dto/create-project-post-industry.dto';
+import { ProjectPostIndustryResponseDto } from './dto/project-post-industry-response.dto';
+import { UpdateProjectPostIndustryDto } from './dto/update-project-post-industry.dto';
 
 @Injectable()
-export class ProjectPostCategoryService {
+export class ProjectPostIndustryService {
   constructor(
     private readonly prisma: PrismaService,
     private readonly prismaErrorService: PrismaErrorService,
     private readonly eventBusService: EventBusService,
   ) {}
 
-  async findAll(): Promise<ProjectPostCategoryResponseDto[]> {
-    const records = await this.prisma.projectPostCategory.findMany({
+  async findAll(): Promise<ProjectPostIndustryResponseDto[]> {
+    const records = await this.prisma.projectPostIndustry.findMany({
       orderBy: [{ id: 'asc' }],
     });
 
     return records.map((record) => this.toDto(record));
   }
 
-  async findById(id: string): Promise<ProjectPostCategoryResponseDto> {
+  async findById(id: string): Promise<ProjectPostIndustryResponseDto> {
     const parsedId = parseInt(id, 10);
     if (Number.isNaN(parsedId)) {
-      throw new NotFoundException(`Category not found for id ${id}.`);
+      throw new NotFoundException(`Industry not found for id ${id}.`);
     }
 
-    const record = await this.prisma.projectPostCategory.findFirst({
+    const record = await this.prisma.projectPostIndustry.findFirst({
       where: {
         id: BigInt(parsedId),
       },
     });
 
     if (!record) {
-      throw new NotFoundException(`Category not found for id ${id}.`);
+      throw new NotFoundException(`Industry not found for id ${id}.`);
     }
 
     return this.toDto(record);
   }
 
   async create(
-    dto: CreateProjectPostCategoryDto,
+    dto: CreateProjectPostIndustryDto,
     userId: number,
-  ): Promise<ProjectPostCategoryResponseDto> {
+  ): Promise<ProjectPostIndustryResponseDto> {
     try {
-      const existing = await this.prisma.projectPostCategory.findFirst({
+      const existing = await this.prisma.projectPostIndustry.findFirst({
         where: {
           name: dto.name,
         },
@@ -65,11 +65,11 @@ export class ProjectPostCategoryService {
 
       if (existing) {
         throw new ConflictException(
-          `Project showcase post category already exists for name ${dto.name}.`,
+          `Project showcase post industry already exists for name ${dto.name}.`,
         );
       }
 
-      const created = await this.prisma.projectPostCategory.create({
+      const created = await this.prisma.projectPostIndustry.create({
         data: {
           name: dto.name,
         },
@@ -78,7 +78,7 @@ export class ProjectPostCategoryService {
       await publishMetadataEvent(
         this.eventBusService,
         'PROJECT_METADATA_CREATE',
-        PROJECT_METADATA_RESOURCE.PROJECT_POST_CATEGORY,
+        PROJECT_METADATA_RESOURCE.PROJECT_POST_INDUSTRY,
         String(created.id),
         created,
         userId,
@@ -86,31 +86,31 @@ export class ProjectPostCategoryService {
 
       return this.toDto(created);
     } catch (error) {
-      this.handleError(error, `create project showcase post category ${dto.name}`);
+      this.handleError(error, `create project showcase post industry ${dto.name}`);
     }
   }
 
   async update(
     id: string,
-    dto: UpdateProjectPostCategoryDto,
+    dto: UpdateProjectPostIndustryDto,
     userId: number,
-  ): Promise<ProjectPostCategoryResponseDto> {
+  ): Promise<ProjectPostIndustryResponseDto> {
     const parsedId = parseInt(id, 10);
     if (Number.isNaN(parsedId)) {
-      throw new NotFoundException(`Category not found for id ${id}.`);
+      throw new NotFoundException(`Industry not found for id ${id}.`);
     }
 
-    const existing = await this.prisma.projectPostCategory.findFirst({
+    const existing = await this.prisma.projectPostIndustry.findFirst({
       where: {
         id: BigInt(parsedId),
       },
     });
 
     if (!existing) {
-      throw new NotFoundException(`Category not found for id ${id}.`);
+      throw new NotFoundException(`Industry not found for id ${id}.`);
     }
 
-    const updated = await this.prisma.projectPostCategory.update({
+    const updated = await this.prisma.projectPostIndustry.update({
       where: { id: BigInt(parsedId) },
       data: {
         ...(typeof dto.name === 'undefined' ? {} : { name: dto.name }),
@@ -120,7 +120,7 @@ export class ProjectPostCategoryService {
     await publishMetadataEvent(
       this.eventBusService,
       'PROJECT_METADATA_UPDATE',
-      PROJECT_METADATA_RESOURCE.PROJECT_POST_CATEGORY,
+      PROJECT_METADATA_RESOURCE.PROJECT_POST_INDUSTRY,
       String(updated.id),
       updated,
       userId,
@@ -132,10 +132,10 @@ export class ProjectPostCategoryService {
   async delete(id: string, userId: number): Promise<void> {
     const parsedId = parseInt(id, 10);
     if (Number.isNaN(parsedId)) {
-      throw new NotFoundException(`Category not found for id ${id}.`);
+      throw new NotFoundException(`Industry not found for id ${id}.`);
     }
 
-    const existing = await this.prisma.projectPostCategory.findFirst({
+    const existing = await this.prisma.projectPostIndustry.findFirst({
       where: {
         id: BigInt(parsedId),
       },
@@ -143,24 +143,24 @@ export class ProjectPostCategoryService {
     });
 
     if (!existing) {
-      throw new NotFoundException(`Category not found for id ${id}.`);
+      throw new NotFoundException(`Industry not found for id ${id}.`);
     }
 
-    await this.prisma.projectPostCategory.delete({
+    await this.prisma.projectPostIndustry.delete({
       where: { id: BigInt(parsedId) },
     });
 
     await publishMetadataEvent(
       this.eventBusService,
       'PROJECT_METADATA_DELETE',
-      PROJECT_METADATA_RESOURCE.PROJECT_POST_CATEGORY,
+      PROJECT_METADATA_RESOURCE.PROJECT_POST_INDUSTRY,
       String(parsedId),
       { id: parsedId },
       userId,
     );
   }
 
-  private toDto(record: ProjectPostCategory): ProjectPostCategoryResponseDto {
+  private toDto(record: ProjectPostIndustry): ProjectPostIndustryResponseDto {
     return {
       id: String(record.id),
       name: record.name,
