+++ src/api/project-showcase-post/project-showcase-post.module.ts
@@ -2,11 +2,23 @@ import { Module } from '@nestjs/common';
 import { GlobalProvidersModule } from 'src/shared/modules/global/globalProviders.module';
 import { ProjectShowcasePostController } from './project-showcase-post.controller';
 import { ProjectShowcasePostService } from './project-showcase-post.service';
+import { ProjectPostCategoryController } from './project-post-category/project-post-category.controller';
+import { ProjectPostCategoryService } from './project-post-category/project-post-category.service';
+import { ProjectPostIndustryController } from './project-post-industry/project-post-industry.controller';
+import { ProjectPostIndustryService } from './project-post-industry/project-post-industry.service';
 
 @Module({
   imports: [GlobalProvidersModule],
-  controllers: [ProjectShowcasePostController],
-  providers: [ProjectShowcasePostService],
+  controllers: [
+    ProjectShowcasePostController,
+    ProjectPostCategoryController,
+    ProjectPostIndustryController,
+  ],
+  providers: [
+    ProjectShowcasePostService,
+    ProjectPostCategoryService,
+    ProjectPostIndustryService,
+  ],
   exports: [ProjectShowcasePostService],
 })
 export class ProjectShowcasePostModule {}
